+++ chia/cmds/init.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import click
 
-from chia.cmds.util import ChiaCliContext
+from chia.cmds.cmd_classes import ChiaCliContext
 
 
 @click.command("init", help="Create or migrate the configuration")
