+++ chia/cmds/stop.py
@@ -7,7 +7,7 @@ from typing import Any
 
 import click
 
-from chia.cmds.util import ChiaCliContext
+from chia.cmds.cmd_classes import ChiaCliContext
 from chia.util.config import load_config
 from chia.util.service_groups import all_groups, services_for_groups
